Added new rule: LocalExpectationRule
This rule verifies that the expected value of an assert is produced locally.
What this means is that it should not be the output of a method call or come
from a field on a class or an object.

Currently the rule allows the expected value to be calculated within the
test method.

The rule *does not* check file system or database access!

Fixes #9